create storePostsForChannel local action
